dont enable darkreader when flag is on, use wrapped components in other compass-components
