+++ packages/compass-components/src/components/checkbox.tsx
@@ -1,12 +1,7 @@
 import React from 'react';
-import LeafyGreenCheckbox from '@leafygreen-ui/checkbox';
-import { css } from '@leafygreen-ui/emotion';
 
+import { Checkbox as LeafyGreenCheckbox } from './leafygreen';
 import { Theme, useTheme } from '../hooks/use-theme';
-
-const checkboxOverrideStyles = css({
-  fontWeight: 'bold',
-});
 
 function Checkbox(
   props: React.ComponentProps<typeof LeafyGreenCheckbox>
@@ -14,11 +9,7 @@ function Checkbox(
   const theme = useTheme();
 
   return (
-    <LeafyGreenCheckbox
-      className={checkboxOverrideStyles}
-      darkMode={theme?.theme === Theme.Dark}
-      {...props}
-    />
+    <LeafyGreenCheckbox darkMode={theme?.theme === Theme.Dark} {...props} />
   );
 }
