+++ packages/compass-components/src/components/confirmation-modal.tsx
@@ -1,7 +1,8 @@
 import React, { useEffect } from 'react';
-import LeafyGreenConfirmationModal from '@leafygreen-ui/confirmation-modal';
 import { createLoggerAndTelemetry } from '@mongodb-js/compass-logging';
 const { track } = createLoggerAndTelemetry('COMPASS-UI');
+
+import { ConfirmationModal as LeafyGreenConfirmationModal } from './leafygreen';
 
 function ConfirmationModal({
   trackingId,
