+++ packages/compass-components/src/index.ts
@@ -22,13 +22,8 @@ import { SpinLoader } from './components/spin-loader';
 import { ResizeHandle, ResizeDirection } from './components/resize-handle';
 import Accordion from './components/accordion';
 export { FavoriteIcon } from './components/icons/favorite-icon';
-export {
-  Variant as BadgeVariant,
-} from '@leafygreen-ui/badge';
-export { Footer } from '@leafygreen-ui/modal';
-export {
-  Variant as BannerVariant,
-} from '@leafygreen-ui/banner';
+export { Variant as BadgeVariant } from '@leafygreen-ui/badge';
+export { Variant as BannerVariant } from '@leafygreen-ui/banner';
 export {
   Size as ButtonSize,
   Variant as ButtonVariant,
@@ -42,12 +37,8 @@ export * as compassFontSizes from './compass-font-sizes';
 export * as compassUIColors from './compass-ui-colors';
 export { default as Portal } from '@leafygreen-ui/portal';
 export { Size as RadioBoxSize } from '@leafygreen-ui/radio-box-group';
-export {
-  Size as SelectSize,
-} from '@leafygreen-ui/select';
-export {
-  Variant as ToastVariant,
-} from '@leafygreen-ui/toast';
+export { Size as SelectSize } from '@leafygreen-ui/select';
+export { Variant as ToastVariant } from '@leafygreen-ui/toast';
 
 export { useToast, ToastArea } from './hooks/use-toast';
